Move dev stack sourcemap resolution to server
